Move 'lift_coefficients_to_xfield' to stark-brainfuck
This function is only used in brainfuck-stark. Move this function there
in case we wish to refactor this STARK engine in the future. It has no
practical use elsewhere.
