+++ stark-brainfuck/src/processor_table.rs
@@ -1,13 +1,13 @@
 use num_traits::{One, Zero};
 use std::convert::TryInto;
 
-use twenty_first::shared_math::b_field_element as bfe;
 use twenty_first::shared_math::b_field_element::BFieldElement;
 use twenty_first::shared_math::mpolynomial::MPolynomial;
 use twenty_first::shared_math::other;
 use twenty_first::shared_math::traits::FiniteField;
 use twenty_first::shared_math::x_field_element::XFieldElement;
 
+use crate::lift_coefficients_to_xfield;
 use crate::stark::{EXTENSION_CHALLENGE_COUNT, PERMUTATION_ARGUMENTS_COUNT};
 use crate::table::{Table, TableMoreTrait, TableTrait};
 use crate::vm::instruction_zerofier;
@@ -511,7 +511,7 @@ impl TableTrait for ProcessorTable {
 
         let mut polynomials: Vec<MPolynomial<XFieldElement>> = b_field_polynomials
             .iter()
-            .map(bfe::lift_coefficients_to_xfield)
+            .map(lift_coefficients_to_xfield)
             .collect();
 
         // extension AIR polynomials
